TelegramTransport: Adjust connectToHost signature, cleanup

+++ TelegramQt/CTcpTransport.cpp
@@ -66,7 +66,7 @@ QString CTcpTransport::remoteAddress() const
     return m_socket ? m_socket->peerAddress().toString() : QString();
 }
 
-void CTcpTransport::connectToHost(const QString &ipAddress, quint32 port)
+void CTcpTransport::connectToHost(const QString &ipAddress, quint16 port)
 {
     qCDebug(c_loggingTcpTransport) << this << __func__ << ipAddress << port;
     m_socket->connectToHost(ipAddress, port);
